Move _qbb and _cbb to core/util/math
