+++ bokehjs/src/lib/core/util/math.ts
@@ -173,3 +173,93 @@ export function eval_poly(poly: Poly, x: number): number {
   }
   return y
 }
+
+/**
+ * Formula from: http://pomax.nihongoresources.com/pages/bezier/
+ *
+ * if segment is quadratic bezier do:
+ *   for both directions do:
+ *     if control between start and end, compute linear bounding box
+ *     otherwise, compute
+ *       bound = u(1-t)^2 + 2v(1-t)t + wt^2
+ *         (with t = ((u-v) / (u-2v+w)), with {u = start, v = control, w = end})
+ *       if control precedes start, min = bound, otherwise max = bound
+ */
+export function qbb(u: number, v: number, w: number): [number, number] {
+  if (v == (u + w)/2)
+    return [u, w]
+  else {
+    const t = (u - v) / ((u - (2*v)) + w)
+    const bd = (u*(1 - t)**2) + (2*v*(1 - t)*t) + (w*t**2)
+    return [Math.min(u, w, bd), Math.max(u, w, bd)]
+  }
+}
+
+// algorithm adapted from http://stackoverflow.com/a/14429749/3406693
+export function cbb(
+    x0: number, y0: number,
+    x1: number, y1: number,
+    x2: number, y2: number,
+    x3: number, y3: number): [number, number, number, number] {
+  const tvalues: number[] = []
+  const bounds: [number[], number[]] = [[], []]
+
+  for (let i = 0; i <= 2; i++) {
+    let a, b, c
+    if (i === 0) {
+      b = ((6 * x0) - (12 * x1)) + (6 * x2)
+      a = (((-3 * x0) + (9 * x1)) - (9 * x2)) + (3 * x3)
+      c = (3 * x1) - (3 * x0)
+    } else {
+      b = ((6 * y0) - (12 * y1)) + (6 * y2)
+      a = (((-3 * y0) + (9 * y1)) - (9 * y2)) + (3 * y3)
+      c = (3 * y1) - (3 * y0)
+    }
+
+    if (Math.abs(a) < 1e-12) { // Numerical robustness
+      if (Math.abs(b) < 1e-12) // Numerical robustness
+        continue
+      const t = -c / b
+      if (0 < t && t < 1)
+        tvalues.push(t)
+      continue
+    }
+
+    const b2ac = (b * b) - (4 * c * a)
+    const sqrtb2ac = Math.sqrt(b2ac)
+
+    if (b2ac < 0)
+      continue
+
+    const t1 = (-b + sqrtb2ac) / (2 * a)
+    if (0 < t1 && t1 < 1)
+      tvalues.push(t1)
+
+    const t2 = (-b - sqrtb2ac) / (2 * a)
+    if (0 < t2 && t2 < 1)
+      tvalues.push(t2)
+  }
+
+  let j = tvalues.length
+  const jlen = j
+  while (j--) {
+    const t = tvalues[j]
+    const mt = 1 - t
+    const x = (mt * mt * mt * x0) + (3 * mt * mt * t * x1) + (3 * mt * t * t * x2) + (t * t * t * x3)
+    bounds[0][j] = x
+    const y = (mt * mt * mt * y0) + (3 * mt * mt * t * y1) + (3 * mt * t * t * y2) + (t * t * t * y3)
+    bounds[1][j] = y
+  }
+
+  bounds[0][jlen] = x0
+  bounds[1][jlen] = y0
+  bounds[0][jlen + 1] = x3
+  bounds[1][jlen + 1] = y3
+
+  return [
+    Math.min(...bounds[0]),
+    Math.max(...bounds[1]),
+    Math.max(...bounds[0]),
+    Math.min(...bounds[1]),
+  ]
+}
