+++ bokehjs/src/lib/models/glyphs/quadratic.ts
@@ -6,27 +6,8 @@ import {inplace} from "core/util/projections"
 import {Context2d} from "core/util/canvas"
 import {Glyph, GlyphView, GlyphData} from "./glyph"
 import {generic_line_vector_legend} from "./utils"
+import {qbb} from "core/util/math"
 import * as p from "core/properties"
-
-// Formula from: http://pomax.nihongoresources.com/pages/bezier/
-//
-// if segment is quadratic bezier do:
-//   for both directions do:
-//     if control between start and end, compute linear bounding box
-//     otherwise, compute
-//       bound = u(1-t)^2 + 2v(1-t)t + wt^2
-//         (with t = ((u-v) / (u-2v+w)), with {u = start, v = control, w = end})
-//       if control precedes start, min = bound, otherwise max = bound
-
-function _qbb(u: number, v: number, w: number): [number, number] {
-  if (v == (u + w)/2)
-    return [u, w]
-  else {
-    const t = (u - v) / ((u - (2*v)) + w)
-    const bd = (u*(1 - t)**2) + (2*v*(1 - t)*t) + (w*t**2)
-    return [Math.min(u, w, bd), Math.max(u, w, bd)]
-  }
-}
 
 export type QuadraticData = GlyphData & p.UniformsOf<Quadratic.Mixins> & {
   _x0: FloatArray
@@ -69,8 +50,8 @@ export class QuadraticView extends GlyphView {
       if (!isFinite(x0_i + x1_i + y0_i + y1_i + cx_i + cy_i))
         index.add_empty()
       else {
-        const [x0, x1] = _qbb(x0_i, cx_i, x1_i)
-        const [y0, y1] = _qbb(y0_i, cy_i, y1_i)
+        const [x0, x1] = qbb(x0_i, cx_i, x1_i)
+        const [y0, y1] = qbb(y0_i, cy_i, y1_i)
 
         index.add_rect(x0, y0, x1, y1)
       }
